Init all members in the constructor
On ubuntu 18.04 pythia.8 crashes due to the  non-initialized variables:
the pythia/include/Pythia8/Info.h class does not initialize most of its members, particularly, in my debug session the isVMDstateAEvent
appears to be set to TRUE and is never set to other value.
Then in the PartonLevel.cc:249 it does
if (infoPtr->isVMDstateA()) tmpBeamA = beamVMDAPtr; // infoPtr->isVMDstateA() returns this isVMDstateAEvent...
but the beamVMDAPtr (which points to the Pythia::beamBDA data member) contains lot of non-initialized pointers,
since its initialization in the Pythia.cc:1018 is conditioned by

if (doVMDsideA)
      beamVMDA.init( 111,  0.5 * eCM, 0.5 * eCM, 0., &info, settings, ...
which is not called since  doVMDsideA is assigned to false.

This patch inits all previously non-initialized members of Init to 0, false etc.